MINOR: Cleanup Java Test Code in Production Classpath
